requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-annoying==0.10.6
django-appconf==1.0.4
django-compressor==2.4
django-extensions==2.2.9
future==0.18.2
idna==2.9
lxml==4.5.0
psycopg2==2.8.5
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
